refactor: adapt old classes to the newest
